+++ app/build.gradle.kts
@@ -54,7 +54,6 @@ dependencies {
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.lifecycle.runtime.ktx)
     implementation(libs.androidx.activity.compose)
-    implementation(libs.gson)
     implementation(libs.koin.android)
     implementation(libs.koin.androidx.compose)
     implementation(libs.retrofit)
