+++ app/src/main/AndroidManifest.xml
@@ -8,14 +8,14 @@
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
-        android:name=".UsersApplication"
+        android:name=".UserApplication"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/Theme.StateGenie"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name="com.arindom.stategenie.presentation.UserListActivity"
             android:exported="true"
             android:theme="@style/Theme.StateGenie">
             <intent-filter>
