+++ app/src/main/java/com/arindom/stategenie/data/model/UserListResponse.kt
@@ -1,10 +1,6 @@
 package com.arindom.stategenie.data.model
 
-data class UserListResponse(
-    val data: List<UserResponse>
-)
-
-data class UserResponse(
+data class User(
     val login: String,
     val avatar_url: String
 )
