+++ app/src/main/java/com/arindom/stategenie/data/network/UserListService.kt
@@ -1,6 +1,6 @@
-package com.arindom.stategenie.data
+package com.arindom.stategenie.data.network
 
-import com.arindom.stategenie.data.model.UserResponse
+import com.arindom.stategenie.data.model.User
 import retrofit2.http.GET
 import retrofit2.http.Query
 
@@ -9,5 +9,5 @@ interface UserListService {
     @GET("users")
     suspend fun getGithubUserList(
         @Query("since") since: String
-    ): List<UserResponse>
+    ): List<User>
 }
