+++ app/src/main/java/com/arindom/stategenie/presentation/UserListActivity.kt
@@ -14,41 +14,28 @@
  * limitations under the License.
  */
 
-package com.arindom.stategenie
+package com.arindom.stategenie.presentation
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.material.CircularProgressIndicator
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Surface
 import androidx.compose.material.TabRowDefaults.Divider
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import com.arindom.stategenie.presentation.UserViewModel
-import com.arindom.stategenie.presentation.`UsersUiState$Generated`
 import com.arindom.stategenie.presentation.widgets.ErrorWidget
 import com.arindom.stategenie.presentation.widgets.ProgressWidget
 import com.arindom.stategenie.presentation.widgets.UserInfo
 import com.arindom.stategenie.ui.theme.StateGenieTheme
-import org.koin.android.ext.android.inject
 import org.koin.androidx.compose.get
-import org.koin.androidx.viewmodel.ext.android.viewModel
 
-class MainActivity : ComponentActivity() {
+class UserListActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
@@ -70,7 +57,7 @@ fun UserListWidget(
             val users = (content as `UsersUiState$Generated`.SuccessState).data
             LazyColumn {
                 items(users) {
-                    UserInfo(modifier = Modifier, userResponse = it)
+                    UserInfo(modifier = Modifier, user = it)
                     Divider(color = Color.LightGray)
                 }
             }
