+++ app/src/main/java/com/arindom/stategenie/presentation/UsersUiState.kt
@@ -1,18 +1,9 @@
 package com.arindom.stategenie.presentation
 
 import com.arindom.stategenie.annotations.StateGenie
-import com.arindom.stategenie.data.model.UserResponse
+import com.arindom.stategenie.data.model.User
 
 @StateGenie(
     rootName = "UsersUiState"
 )
-interface UsersUiState: BaseUIiState<List<UserResponse>>
-
-data class UserList(
-    val users: List<User> = emptyList()
-)
-
-data class User(
-    val name: String,
-    val email: String
-)
+interface UsersUiState: BaseUIiState<List<User>>
